% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification-package.R
\docType{package}
\name{classification-package}
\alias{classification}
\alias{classification-package}
\title{classification: Classify Two-Class Data}
\description{
Provide tools for classifying two-class data.
}
\author{
\strong{Maintainer}: Adam Shen \email{adam.shen@carleton.ca}

}
\keyword{internal}
