% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify}
\alias{classify}
\title{Create classify}
\usage{
classify(df, target, method)
}
\arguments{
\item{df}{input data - predictors}

\item{target}{input data - target}
}
\value{
classify
}
\description{
Create classify
}
