% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Confusion Matrix and Metrics}
\usage{
evaluate_model(y_true, y_pred, cutoff = 0.5)
}
\arguments{
\item{y_true}{True binary response values.}

\item{y_pred}{Predicted probabilities.}

\item{cutoff}{Cut-off value for classification (default = 0.5).}
}
\value{
A list containing confusion matrix and various metrics.
}
\description{
Confusion Matrix and Metrics
}
