% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_tibble.R
\name{matrix_to_tibble}
\alias{matrix_to_tibble}
\title{Coerce a matrix to a tibble}
\usage{
matrix_to_tibble(x)
}
\arguments{
\item{x}{A matrix.}
}
\description{
Convenience function for coercing a \eqn{d \times n} matrix into a tibble with \eqn{n} rows and
\eqn{d} columns, where \eqn{n} is the number of observations and \eqn{d} is the number of
variables.
}
