% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_mle.R
\name{mean_mle}
\alias{mean_mle}
\title{Compute maximum likelihood estimate of population mean}
\usage{
mean_mle(x)
}
\arguments{
\item{x}{A matrix of dimension \eqn{d} by \eqn{n}, where \eqn{d} is the number of dimensions and
\eqn{n} is the sample size.}
}
\value{
A matrix of dimension \eqn{d} by \eqn{1}.
}
\description{
For data from a multivariate normal distribution, the maximum likelihood estimate of the
population mean is computed as the sample means of the individual dimensions.
}
