% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvtnorm.R
\name{rmvtnorm}
\alias{rmvtnorm}
\title{Sample from the (approximate) multivariate normal distribution}
\usage{
rmvtnorm(n, mean, cov)
}
\arguments{
\item{n}{The sample size.}

\item{mean}{A \eqn{d \times 1} matrix containing the means of the variables.}

\item{cov}{A \eqn{d \times d} positive-definite, symmetric matrix specifying the covariance
matrix of the variables.}
}
\value{
A matrix with \code{d} rows and \code{n} columns.
}
\description{
Sample from the (approximate) multivariate normal distribution
}
\details{
Consider a random sample of size 12 of Uniform(0,1) random variables. Let \eqn{X} be the
sum of the sample elements. Then \eqn{X} has an approximate \eqn{N(0,1)} distribution.

Let \eqn{\mathbf{Z}} be a \eqn{d \times n} matrix with distribution
\eqn{N_{d}(\mathbf{0},\,\mathbb{I})}. Let \eqn{\mathbf{X}} be a \eqn{d \times n} random matrix
with distribution \eqn{N_{d}(\bm{\mu},\,\mathbf{\Sigma})}. If \eqn{\mathbf{P}} and
\eqn{\mathbf{D}} are the eigenvector and eigenvalue matrices, respectively, then

\deqn{\mathbf{X} \,=\, \mathbf{PD}^{\frac{1}{2}}\mathbf{Z} \,+\, \bm{\mu}\mathbf{1}'
\quad\sim\quad N_{d}(\bm{\mu},\, \mathbf{\Sigma}),}

where \eqn{\mathbf{1}} is a column vector of ones.
}
\examples{
mu <- matrix(c(1, 3))
sigma <- matrix(c(3, 1, 1, 2), 2, 2)
rmvtnorm(n = 5, mean = mu, cov = sigma)
}
