% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seeds}
\alias{seeds}
\title{Measurements of geometrical properties of wheat kernels}
\format{
A tibble with 140 rows and 7 variables:
\describe{
  \item{area}{Area of the kernel}
  \item{peri}{Perimeter of the kernel}
  \item{length_k}{Length of the kernel}
  \item{width_k}{Width of the kernel}
  \item{asymmetry}{Asymmetry coefficient}
  \item{length_g}{Length of the kernel groove}
  \item{type}{Class of the seed (\code{class1} or \code{class2})}
}
}
\source{
M. Charytanowicz, J. Niewczas, P. Kulczycki, P. Kowalski, and S. Lukasik. (2012) seeds.
UCI Machine Learning Repository. \url{https://archive-beta.ics.uci.edu/ml/datasets/seeds}
}
\usage{
seeds
}
\description{
A dataset containing measurements of six geometrical properties for 140 wheat kernels of two
different types. The 140 observations are composed of 70 observations from seeds of \code{class1}
and 70 observations from seeds of \code{class2}.
}
\keyword{datasets}
