% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{summary.classify}
\alias{summary.classify}
\title{Summary of classify
Print summary of test and train data used for classification}
\usage{
\method{summary}{classify}(x)
}
\arguments{
\item{x}{classify}
}
\description{
Summary of classify
Print summary of test and train data used for classification
}
