% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvtnorm.R
\name{rmvtnorm}
\alias{rmvtnorm}
\title{Sample from the (approximate) multivariate normal distribution}
\usage{
rmvtnorm(n, mean, cov)
}
\arguments{
\item{n}{The sample size.}

\item{mean}{A \eqn{d \times 1} matrix containing the means of the variables.}

\item{cov}{A \eqn{d \times d} positive-definite, symmetric matrix specifying the covariance
matrix of the variables.}
}
\value{
A matrix with \code{d} rows and \code{n} columns.
}
\description{
Sample from the (approximate) multivariate normal distribution
}
\details{
Consider a random sample of size 12 of Uniform(0,1) random variables. Let \eqn{X} be the
sum of the sample elements. Then \eqn{X} has an approximate \eqn{N(0,1)} distribution.

Let \eqn{Z} be a \eqn{d * n} matrix with distribution \eqn{N_d(0, 1_d)}. Let \eqn{X} be a
\eqn{d * n} random matrix with distribution \eqn{N_d(\mu, \Sigma)}. If \eqn{P} and \eqn{D} are
the eigenvector and eigenvalue matrices of \eqn{\Sigma}, respectively, then

\deqn{X = PD^{1/2}Z + \mu1'}

has a \eqn{N_d(\mu, \Sigma)} distribution, where \eqn{1} is a column vector of ones.
}
\examples{
mu <- matrix(c(1, 3))
sigma <- matrix(c(3, 1, 1, 2), 2, 2)
rmvtnorm(n = 5, mean = mu, cov = sigma)
}
